Workarounds for g++ 4.9 series.
--HG--
branch : multiweight

+++ include/Rivet/AnalysisHandler.hh
@@ -68,7 +68,7 @@ namespace Rivet {
         return _eventCounter->sumW();
     }
 
-    int numWeights() const {
+    size_t numWeights() const {
         return _weightNames.size();
     }
